refact : 랜덤 셋팅 로직 분리
